requirements.txt
pytelegrambotapi==4.19.1
python-dotenv==1.0.0